hola

+++ Inc/lcd.h
@@ -39,7 +39,7 @@ void LCD_Put_Char(uint8_t c);
 void LCD_Init(void);
 void LCD_Set_Cursor(uint8_t line, uint8_t column);
 void LCD_Put_Str(char * str);
-void LCD_Put_Num(int16_t num);
+void LCD_Put_Num(int32_t num);
 char LCD_Busy(void);
 void LCD_Pulse_EN(void);
